fix: Add missing entity properties and fix type mismatches
Entity Model Updates:
- RelatedCase: Add CourtId (int?), Court navigation, CaseNumber (long), CaseYear (int)
- RequestClassification: Add ClassificationId (int), Classification navigation
- CaseRegistrationRequest: Add CaseTypeId (int?), PrimaryMobile, SecondaryMobile, Email properties
- IUnitOfWork: Add ClearChangeTracker() method for change tracker management

ViewModel Updates:
- RelatedCaseVM: Update CaseNumber to long to match entity

These changes resolve build errors in business logic services that were referencing undefined entity properties.

+++ src/Backend/BOG.VM/CaseRegistration/RelatedCaseVM.cs
@@ -10,7 +10,7 @@ public class RelatedCaseVM
     public int CaseRegistrationRequestId { get; set; }
     public int? CourtId { get; set; }
     public string? CourtName { get; set; } // For display purposes (from Court.NameAr)
-    public int CaseNumber { get; set; }
+    public long CaseNumber { get; set; }
     public int CaseYear { get; set; }
     public DateTime CreatedDate { get; set; }
     public DateTime ModifiedDate { get; set; }
